testsuite: make fake imp chdir back to trash-directory
Problem: The fake IMP in t/job-exec/imp.sh writes out files to current
working directory, but the job-exec system may not change the working
directory of the IMP to the working directory of the job, since this
is not always possible.

Add

 cd $SHARNESS_TRASH_DIRECTORY

at the top of imp.sh so that the test IMP will have the expected
working directory.