job-exec: do not change working directory of multiuser jobs
Problem: The job-exec module attempts to launch job shells or the
IMP in the working directory of the job. This won't work for
multiuser jobs, because the instance owner may not have access to the
working directory of the job. This results in an error like:

 Could not change dir to /home/user1: Permission denied. Going to
 /tmp instead

Always start the IMP in the root directory, "/". This avoids the
unnecessary error mentioned above, and also avoids starting a possibly
setuid process in a world-writeable directory like "/tmp".