docs: add explicit link to man pages section
Problem: when referencing the flux-core docs via intersphinx, it uses
the top section title of `index.rst` "Welcome to flux-core's
documentation".  For the meta flux-docs RTD project, we want a link that
says "flux-core Man Pages".

Solution: Rename the subsection to "flux-core Man Pages" and give it an
explicit link so that we can reference it with intersphinx from
flux-docs.  While we are at it, remove the now redundant "Manual Pages"
caption and rename the title to "Documentation for flux-core".