testsuite: add dmesg-grep.py test script
Problem: The construct used in tests "flux dmesg | grep PATTERN"
is sometimes racy, e.g. if the log message is not sent from broker
rank 0, or the test script is waiting for a log message that may be
produced a short time after the event which triggered it.

To avoid the need to run flux-dmesg in a loop, introduce a test
script t/scripts/dmesg-grep.py which follows the dmesg log watching
for regex pattern. If no line matches after a timeout, then the script
exits with an error.