python: allow future.then() callbacks to take varargs
Problem: The current limit of a single argument to the future.then()
continuation callback results in poor ergonomics in Python code.
Users of future.then() end up creating dicts on the fly or using other
shenanigans to pass multiple arguments to the callback.

Allow future.then() to be called with varags and keyword args. Store
the variable argument list in future.then_args, and the keyword
args in future.then_kwargs, and pass these directly to the provided
callback. This gives the Python developer a more familiar interface
when using Future callback functions.