build: libsubprocess: fix bad link order in unit tests
Problem: Unit tests in libsubprocess link against system libflux-core
when flux-core is built --with-flux-security. Apparently this can
occur when libflux-core.la is not last on the LDADD line. libtool
gets confused and adds an rpath to /usr/lib64 or the system library
path.

Change the order of libs in LDADD to keep libflux-core.la last, and
make sure libtap is first. This seems to work around the issue for now.