docker: install poison libflux into docker images
Problem: When building flux-core on a system with a different version
installed, we want to ensure that the testsuite run by `make check`
never links against the system libflux-core.so and other libraries,
and does not use the system flux(1) executable. However, the docker
images used by docker-run-checks.sh and in CI tests does not have a
flux-core installed, so problems in this area may go undetected.

Add functionality in travis_run.sh to install "poison" versions
of libflux-core,idset,hostlist,schedutil and optparse, which
are essentially empty libraries that error in the initialization
function. If any part of 'make check' links with any of these system
installed libs, the testsuite will get an immediate error. For good
measure, also install a poison flux executable in /usr/bin that also
errors out.

The libraries are generated in travis_run.sh via a new script
src/test/docker/libflux-poison.sh. This script is built from configure
via libflux-poison.sh.in so that LIBDIR, BINDIR and the various
libflux-* versions can be substituted into the script. This ensures
that the correct DSO version links are installed.

The poison libflux install can be disabled via a new `--no-poison`
option to docker-run-checks.sh.