job-manager: support flux-restart event
Per RFC21, support the flux-restart event in the eventlog.  When
read from the eventlog on restart, roll back from the SCHED state
to PRIORITY state.  When in the SCHED state after a restart, generate
the flux-restart event to regenerate the 'priority' event.

Update job-info to handle the flux-restart event as well.

Fixes #3345