job-manager: differentiate admin vs queue priority
Problem: The job structure in the job-manager stores the priority
in a single variable "job->priority".  This can be confusing as
there are two priorities, the administrative priority and the queue
priority, and both are stored in the same variable.

Solution: Store the admin and queue priorities separately in
job->admin_priority and job->queue_priority appropriately.  Update
sort comparisons to use the queue_priority.  Update tests accordingly.