job-manager: support priority-update event
Problem: When the queue priority changes due to an admin priority
change, there is no notification to the job-info module that
the queue priority has changed.

Solution: Add a new event priority-update event that will indicate
when the queue priority has been changed.  Update job-info to
parse this event in the journal and from the eventlog.

Fixes #3345