shell: ensure TMPDIR exists
Problem: Some sites may set TMPDIR to a per-user directory which may
not exist on compute nodes. Since the Flux broker may use TMPDIR as
a component of rundir, this may cause errors such as the following
for batch jobs:

 broker: create_rundir: No such file or directory

Add a Lua plugin to the shell which will ensure TMPDIR exists if set
in the job's environment.

Fixes #3326