build: bail out on 16 bit ints
Problem: Flux assumes that the C int type is at least 32 bits.
For example, job priorities, defined as a 32 bit unsigned value,
are encoded in JSON objects using the jansson "i" format specifier.

Add a configure time check that the C int type is not size 16.
This seems pretty unlikely to come up but let's just be explicit anyway.

Fixes #3385