job-manager: capture readacted jobspec from submit rpc
Problem: the job manager priority plugin may need access to
the jobspec in the future, but job-manager doesn't have it.

jobspec is now provided in the `job-manager.submit` request.
Retain it in the `struct job` for future use.

Update submit unit test, adding an empty jobspec object to
the test submit request.

Fixes #3336