job-ingest: parse jobspec in mainline
Problem: jobspec is parsed in the validation code and thrown away,
but we want to keep it to pass to the job manager now.

Move jobspec parsing (and capture of JSON parse errors) to the
submit RPC handler.  This is clearer anyway, since before we had
to create a fake future to return textual JSON parse errors in,
and now we just use the existing RPC error path for that.