job-exec: remove cleanup phase
Problem: The job-exec "cleanup" phase is not used nor useful. It
is meant as a placeholder for a job epilog run via the IMP, but it
only runs after *all* ranks have completed, which is not what we
eventually want. This "cleanup" stage of the exec system is therefore
just waste of processing in the module and a couple useless events
in the eventlog.

Remove the cleanup phase. Per rank epilog will be added at a later
time, at which point the exec events can be re-introduced if necessary.