job-manager: send redacted jobspec in sched.alloc
Problem: the scheduler has to look up the jobspec upon
receipt of a sched.alloc request, but the job manager has
it in memory now.

Send redacted jobspec directly to the scheduler.
The scheduler should not require the redacted environment.

This will require an update to RFC 27, but schedulers will
not need to be modified since jobspec is handled by libschedutil.