libschedutil: drop auto-ENOSYS response to alloc/free
Problem: now that a disconnect request is generated when modules
unload, it is no longer important for job manager requests to the
scheduler to receive an ENOSYS response when the scheduler unloads.
The job manager's disconnect handler takes care of clearing state
related to outstanding scheduler requests.

Drop the code in libschedutil for sending ENOSYS responses for
requests that have not yet been sent to the scheduler because they
are waiting on another async operation such as a KVS commit.
We can simply discard those now.

Drop the test code that verifies the ENOSYS responses.