libschedutil: accept jobspec in alloc request
Problem: Now that jobspec (minus environment) is provided by the job
manager in the sched.alloc request, it is no longer necessary for
libschedutil to fetch it from the KVS for the scheduler.

Drop the the code for looking up the jobspec asynchronously and the
queue for maintaining the received order of alloc requests.

The libschedutil API is unchanged.