libeventlog: add eventlogger_commit()
Problem: There is no way to "flush" the current batch of an eventlogger
and asynchronously wait for the KVS commit to complete.

Add a new public function `eventlogger_commit()` which commits the
current batch and returns a future which will be fulfilled when the
kvs transaction has been committed.

Use a composite future so that the caller doesn't have to call
eventlog_batch_complete() manually, instead this is done always via an
"and then" callback.

Refactor and update common code so that the synchronous version,
eventlogger_flush() is now implemented using eventlogger_commit().