libeventlog: add eventlogger_setns()
Problem: The eventlogger code only works for the primary namespace
since it calls flux_kvs_commit() with ns hard-coded to NULL.

Add a namespace member of the eventlogger object, and a new setter
function `eventlogger_setns()` to set a namespace.

Use the currently set namespace when calling flux_kvs_commit(3).