job-exec: log shell/imp output to exec eventlog
Problem: Early job shell or IMP errors are not written to the job
output eventlog since that log is create by and for the shell. Instead
errors on standard output of these programs is currently only logged
to the Flux broker dmesg log, and is lost to the user, resulting in
useless messages like:

 flux-job: task(s) exited with exit code 1
 flux-job: No job output found

when the shell or IMP encounter early errors.

Expose a new `jobinfo_log_output()` function which can be used by
exec implementations to log output, errors, or anything else to the
exec eventlog under a new "log" type event. Use this function in the
subprocess exec implementation to log shell/imp output.

Keep the flux_log() for now as well, since some tests depend on it.