t2402-job-exec-dummy.t: read shell output from flux-job attach
Now that the job shell output is available from the exec.eventlog
and thus flux-job attach, use that method to look for expected output.

This allows the exec system to eventually transition away from logging
this output via flux_log(), and also serves as a test for expected
behavior of `flux job attach` handling of exec.log events.