broker: avoid using the zmq monitor socket
Problem: the zmq socket monitor has a critical bug in 4.1.4,
but we're stuck with that version on centos 7.

Stop using the zmq socket monitor entirely.  Instead, track
connects and disconnects by monitoring message traffic.

Connect is detected when the first message arrives.

Disconnect is detected when a message is sent to a child,
and the send function returns EHOSTUNREACH.  Given the periodic
heartbeat that is distributed from parent to child over the TBON,
worst case, disconnect is detected within a heartbeat interval.

This delay will be addressed in a subsequent commit.