broker: enable heartbeat-synchronized idle detect
Problem:  idle node detection was disabled long ago, but is useful
for debugging broker deadlocks, such as the one just discovered.

First, the subscription topic is "heartbeat", which is incorrect.
Change to "hb'.

Next, set the idle warning so that a message is logged when a broker
has not checked in for 5 heartbeats.

Finally, ensure that broker ranks that have already disconnected
proprerly are not logged as idle.

In the future, this feature will likely need to be reworked to function
more smoothly for a system instance, but for now the old behavior is
useful for debugging.