job-exec: do not log exception to exec.eventlog
Problem: For reasons unknown, the job-exec module attempts to log an
"exception" event to the guest.exec.eventlog, even though exceptions
are already emitted to the main job eventlog. This can cause a race
with the eventlogger, since the job namespace may be removed by the
time the exception event is committed.

Remove this useless event rather than attempting to fix the race.
There is no reason to emit it.