broker: use non-blocking mode on router socket
Problem: the ROUTER socket facing downstreawm TBON children
seems to block occasionally on zmq-4.1.4 (Centos 7).

According to zmq_socket(3), the router socket should never block,
e.g. its "mute state" is "drop".  It seems this is likely a bug
peculiar to this version of zmq, since we don't see it in other
versions.

Add flux_msg_sendzsock_ex() with a "nonblock" argument and use it
with nonblock=true for this socket only.  In this situation, the
function now fails with EAGAIN, instead of blocking the broker
reactor.  The error is logged at LOG_ERR severity.

Fixes #3235