broker: send disconnect notification to parent
Problem: TBON child disconnect detection is potentially delayed
until a heartbeat event fails with EHOSTUNREACH.

Send a keepalive with a special status value just before to tearing
down the parent socket.  Since we set zmq_linger=5 (msec), there should
be enough time in most cases for this message to reach the parent before
the socket is closed.  If it doesn't make it, EHOSTUNREACH on the next
heartbeat triggers disconnect detection.