job-info: fix potential missed state transition
Problem: when reverting a job's state due to a flux-restart event in
the journal, the current job->state value is checked to be be SCHED.  This
check is invalid, as it is the current job state that has been processed.  It
is possible state transitions are still being processed and we have not
yet reached SCHED in the internal state transition engine.

Solution: When ultimately calling add_state_transition(), add a flag indicating
it is a revert, so that appropriate checks can be skipped, and add a state that
it is expecting before the revert.