python: add ResourceSet class
Problem: The python Rlist class only provides support for Rv1 users,
and usage is specific to librlist idiosyncracies.

Introduce a general ResourceSet class which will (hopefully) define
a common interface for consumers of R, preventing the need to update
all callers or break out-of-tree users whenever there is new revision
of the underlying R representation.

The ResourceSet class uses an underlying, version-specific,
implementation, which must be a member of the ResourceSetImplementation
base class.  This abstract base class defines the interface that
a ResourceSet implementation must provide in order to allow the
ResourceSet class to work properly.

Define the Rlist class as the implementation for version 1 of the
ResourceSet class.