flux-resource: drain: list drained targets with no args
Problem: There is no specific command to list drained targets along
with the timestamp provided in the resource.status RPC response.

Change the `flux resource drain` command such that the list of targets
is optional, and display a list of currently drained targets when
a target set to drain is not supplied.