job-manager: do not clear all annotations
Problem: On an alloc cancellation response, the annotations of a job
are always cleared, including user annotations.  This was based on the
assumption that an allocation cancel response meant the job was completed.
Per RFC27, this is not always the case.  In addition, during an interface
teardown annotations are completely cleared, even though a job may be scheduled
again in the future.

Solution: If the job is re-schedulable in the future, only clear scheduler
annotations.  Do not clear all annotations.