job-manager: cancel alloced job if priority low
Problem: In single mode, the job-manager may have at most one
alloc request outstanding with the scheduler.  If a new job is
submitted with a higher urgency, or an existing job is given
a higher urgency, the outstanding alloc is not cancelled.  Therefore,
a lower urgency/priority job is executed before a higher one.

Solution: After a new job submission or change to a job's urgency
is made, make sure that the job at the top of the alloc queue is of lower
priority than the one sent to the scheduler.  If it isn't, cancel
the job and place it back on the queue.  Existing code will eventually
send the new highest priority job to the scheduler.  Update tests in
t/t2234-job-info-list-update.t for this change.

Fixes #3333