codecov: remove ignore of bindings/python/.*.c
Problem: The coverage reports on codecov.io are missing some Python
source files, apparently due to the ignore line:

  - ".*/bindings/python/.*.c"

which appears to ignore any file with a 'c' character in the name
under src/bindings/python (?).

Replace this line with

  - "src/bindings/python/_flux"

to ignore the whole generated bindings directory, which is the intent.