job-info: cache finish event status
Problem: While the job-info module caches whether a job is successful
or not via the success attribute, it does not cache the raw wait status,
which may be useful to consumers of job information. Instead these
consumers are forced to read the job eventlog.

Add a new wait_status member of `struct job` and a corresponding
"waitstatus" attribute which can be used to fetch it. The wait_status
is initialized to -1 and is only valid when set >= 0 from a job
"finish" event.