python: add waitstatus and returncode attributes to JobID
Problem: There is not a convenient way to get the wait status,
exit code, or terminating signal of jobs from instances of the
JobInfo class.

Support the new `waitstatus` attribute in job-info.list.

Introduce a new computed property, `returncode`, which represents
an overall return code for the job. If the job is still active, the
return code is an empty string. For inactive jobs, a return code
is the exit code computed by `os.WEXITSTATUS(waitstatus)` if the
job exited, -SIG if the job was terminated by signal SIG, or the
special value -128 if the job was cancelled before it was started.

This allows all inactive jobs to have a valid return code, and the
return code will be non-zero for any job that did not normally exit
with 0 exit code.

Fixes #3408