flux-jobs: support waitstatus and returncode fields
Problem: JobInfo objects support the new waitstatus and returncode
properties, but these are not supported in flux-jobs since they do
not appear in the fields2attrs mapping.

Add these fields to the fields2attrs mapping in fetch_jobs_flux()
so that they may be specified in flux-jobs output format strings.