t/python: add python3 she-bang lines
Problem: The python tests in the testsuite can't be run from the
commandline directly like other tests in the suite, because they
either are not executable or lack the correct python3 she-bang line.

Fix non-executable test scripts and/or add python3 to the interpreter
lines of these tests.

The environment still needs to be modified to run the Python tests so
that pycotap can be found, e.g.

 PYTHONPATH=$PYTHONPATH:$(pwd)/python/tap ./python/t0001-handle.py