python: security: fix utf-8 handing in sign_wrap_as
Problem: When sign_wrap_as() is used in a non-UTF-8 locale, with
input containing unicode or non-ascii characters, sign_wrap_as()
can throw a Unicode error.

Add errors="surrogateescape" to the payload.encode() so that these
characters are properly escaped in the signed payload and the
Python exception is avoided.