python: fix syntax warnings from Sphinx due to docstrings
Problem: several of the docstrings within the python bindings are not
valid sphinx or google-style docstrings.  This causes a multitude of
sphinx warnings when generating the html docs.  In particular, these
warnings included `unexpected indent`s due to spacing issues or a
multiline description without a proper indent, unexpected ReST headers,
unescaped `*`s, and yaml blocks in markdown format (as opposed to ReST).

Fix all of the various sphinx warnings in one go with various tweaks
throughout the binding's docstrings.

Co-authored-by: Elsa Gonsiorowski <gonsie@llnl.gov>