flux-jobs: add --stats and --stats-only options
Problem: A simple method with nice output does not exist to query
the job statistics provided by the job-info.job-stats RPC.

As a first cut, add `--stats` and `--stats-only` options to
flux-jobs(1) which display a simple text summary of the most
interesting job stats.  The stats line is printed above the jobs
output header with `--stats`, and is the only output if `--stats-only`
is used.

As an added enhancement, `flux jobs --stats-only` will exit with
non-zero exit status when there are no more active jobs. This allows
the following loop to exit naturally after all pending and running
jobs have completed:

 while flux jobs --stats-only; do sleep 2; done

Fixes #3386