flux-keygen: generate cert at specified PATH
Problem:  the main use case for flux-keygen is now generating
keys for a system instance, so it should not write to home
directory by default.

Change the command usage so that a certificate containing both public
and secret keys is saved to PATH.  Issue a warning if PATH is unspecified,
but don't fail so that flux-sched tests continue to work for now.

Drop 'keydir' and FLUX_SEC_DIRECTORY support from the flux command front
end and libflux/conf, and skip auto-generation of keys for in-tree builds.

Update the flux-keygen tests in t0001-basic.t for the new command usage.

Update t0013-config-file.t to start the configured instance using a
configured curve cert.