lua: json_object_to_lua: fix segfault on JSON_NULL
Problem: If a JSON_NULL appears in a json_t object Lua may segfault
since the value is silently ignored and the table assignment could
be left dangling.

Instead of uselessly bailing, set the value to Lua `nil`, effectively
unsetting any table key, but this is better than a segfault.

Fixes #3415