t4000-issues-test-driver: allow scripts to register prereqs
Problem: Scripts run by the t4000-issues-test-driver.t test do not
have a way to register prereqs. For example if a script reuqires
jq(1) it cannot be run via this driver.

Allow scripts to register prereqs to be passed to test_expect_success()
by setting the string

 test-prereqs: PREREQ,...

anywhere in the the script.