libjob: add FLUX_JOB_NR_STATES
Problem: It is difficult to determine the total number of job states
defined in the flux_job_state_t enum, since this enum is defined as
a set of flags instead of a set of monotonically increasing values.
This means the typical tricks of defining a FLUX_JOB_STATE_MAX as
the last enum value do not work. Nor can we rely on the value of the
"final" state FLUX_JOB_STATE_INACTIVE.

Use the simplest solution and add a new #define FLUX_JOB_NR_STATES
which users of libjob can use to determine the total number of job
states which are defined in the enum.