flux-mini: add bulk job submission features
Problem: There is not an easy way to submit copies of the same job to
Flux rapidly. Currently this is multistep process involving generation
of jobspec, then use of the test-only submitbench command or hand-coded
scripts to submit copies of the generated jobspec asynchronously.

Add a set of new bulk job submission capabilies to flux-mini, including
the following new options for `submit`:

 * --cc=IDSET, --bcc=IDSET, which replicates the generated jobspec
   once for each ID in the supplied IDSET. With --cc only, the
   environment variable `FLUX_JOB_CC=ID` is set for each job.

 * --wait, which causes `flux mini submit` to wait until all submitted
   jobs have completed or failed.

 * --watch, which additionally displays output for each submitted job.

 * --progress, which displays a progress bar indicating the percentage
   of jobs that have completed, along with counts of jobs in various
   states.

Also add a new subcommand `bulksubmit`, which works like xargs(1)
or GNU parallel and allows submitting a set of jobs based on one
or more inputs supplied on stdin or the command line.

The bulksubmit command takes all the same options as ``submit``, but
many options can be substituted using replacement strings which
are formatted using string.format() internally. For a full description
see the BULKSUBMIT section of the flux-mini(1) manpage.