github: allow git fetch --tags to fail in actions
Problem: When GitHub actions are run as a result of a tag
being pushed, that tag will already be present in the git clone,
and `git fetch --tags` will fail with a warning about not
overwriting existing tags.

Make `git fetch --tags` best-effort in actions. We don't care if
it fails in the case of running the action directly on a tag, but
we don't want the whole action to fail.