python: fix circular Future reference on exception
Problem: check_future_error() saves the result of sys.exc_info() as a
local variable, which causes a circular reference. While not fatal in
any way, this can delay the garbage collection of objects referenced
in the stack trace, which can in turn cause python programs to get
deadlocked in the Flux reactor (e.g. if a future is not destroyed
after its continuation callback is complete.)

Remove the use of sys.exc_info() and instead reraise exceptions
from the orignal saved 'error'.

Additionally raise the exceptions `from None` so that we don't
generated the exception-within-an-exception error message.

Fixes #3429