python: add ability to check if Flux reactor is active
Problem: There is currently no way for Python process to determine
if the current thread is executing within a Flux reactor. This may
be necessary to enable or disable certain features when the reactor
is active. This value must be global and not per handle instance,
because Python will instantiate a new `Flux` object when in callbacks
using the same C handle. The value must be per-thread because a Python
program may be composed of multiple threads, of which only one may
be currently executing a Flux reactor.

As a stopgap, instantiate a class-wide thread-local storage variable in
flux.core.handle.Flux, which is incremented upon entry to the Flux
reactor (from any handle instance), and decremented upon exit from
the reactor.

Add a reactor_running() class method so that the current thread-global
state can be queried with flux.core.handle.Flux.reactor_running()
or the equivalent.