python: re-throw uncaught exceptions from future callbacks
Problem: Uncaught exceptions in future then() callbacks may print a
message to stderr, but the exception state is lost when re-entering
the Flux reactor, and thus a program may continue running, and when
it does exit the reactor, any error state is lost.

Add a global, per-thread exception state to the flux.core.handle.Flux
class. If the reactor exits with an error, and there is a pending
exception for this thread, then that exception is re-thrown.

The exception state is then set in flux.future.continuation_callbak()
if an exception is thrown during the future's then_cb().