python: re-throw exceptions from watcher/message handlers
Problem: Uncaught exceptions in Flux handle watchers and message
handlers may emit an message to stderr, but they are otherwise
unhandled. Once the reactor is re-entered, the error state is lost
and the program may not realize there was an error when the reactor
finally exits.

Set the global Flux handle exception state when an uncaught exception
is thrown from user callbacks in watchers and message handlers.

This will trigger the re-raise of the exception already in place in
the Flux handle reactor_run() method.