job-manager: pack/unpack priority with 64 bit int
Problem: Internally flux sends priority values as an int when using
jansson pack/unpack routines.  However, priority values have a range
from 0 to 2^32-1, i.e. the range of an unsigned int.  If the value of a
priority exceeds INT_MAX, there can be issues if the value is not
read/unpacked and re-cast into an unsigned int.

Solution: Internally store the priority of a job using a int64_t instead
of a unsigned int.  Pack/unpack the job priority using the 'I' format
specifier (json_int_t, which is a 64 bit int).  Update correspondingly
in job-info.

Fixes #3449