job-manager: allow hold on all pending jobs
Problem: In the job-manager, once a job enters pending, it can no
longer have its urgency changed.  This means that the job hold
mechanism does not work when the job-manager runs in unlimited mode.

Solution: Specifically for the case of a job hold, allow a job to
be canceled and be re-queued into the job-manager's queue.