shell: unbuffer stdout before task.exec callback
Problem: Output written to stdout in the shell task.exec plugins could
be lost since stdout is fully buffered, and the buffer doesn't have a
chance to be flushed before exec(2) is called.

Set stdout to unbuffered before calling the task.exec callbacks to avoid
this problem, and the confusion it may cause.

Fixes #3444