python/job: add RFC21 reference to the event wait/watch docstrings
Problem: the exact events that will appear in the eventlog and their
contents are rigorously documented in RFC21.  It would be good to have a
pointer to that from the docstrings of the `event_wait` and
`event_watch` functions

Solution: add intersphinx support to the sphinx `conf.py` so that we can
reference the contents of other Flux sphinx projects.  Then add a
`.. seealso` section to each functions docstring that references the
RFC21 document with the name `rfc:spec_21`.