connectors/ssh: forward LD_LIBRARY_PATH over ssh when set
Problem: ssh and rsh do not forward environment variables, thus
LD_LIBRARY_PATH is not guaranteed to be set on the remote node where the
flux command is run.  If the flux command is linked against libraries
that can only be found when LD_LIBRARY_PATH is set, then the flux
command will fail to run over ssh.

Solution: Grab the client-side LD_LIBRARY_PATH so that we can manually
forward it.  Then forward it by prepending it to the flux relay
command (e.g., `ssh hostname LD_LIBRARY_PATH=/foo flux relay`).