libflux: plugin: add flux_plugin_get/set_flags
Problem: The dlopen(3) flags used in flux_plugin_load_dso(3) are
hardcoded to

  RTLD_LAZY | RTLD_LOCAL | FLUX_DEEPBIND

However, some plugin loaders may want to change one or more of these
flags, e.g. RTLD_NOW instead if RTLD_LOCAL so that function binding
is immediate and caught as an error in dlopen(3) instead of a later
crash if symbol resolution fails.

Encapsulate the dlopen flags into an internal flags variable so that
a getter and setter for dlopen flags can be introduced in the API, and
expose a set of FLUX_PLUGIN_RTLD_* flags that, for now, correspond
1:1 to the glibc RTLD_* flags.

This allows a caller to adjust flags before calling
flux_plugin_load_dso(3), e.g.:

 flux_plugin_set_flags (p, FLUX_PLUGIN_RTLD_NOW);
 flux_plugin_load_dso (p, path);

should fail if the DSO calls any functions that are not immediately
resolved.