broker: load modules with RTLD_GLOBAL
Problem: Currently all broker modules are loaded with RTLD_LOCAL,
which is a valid default since modules are not meant to be sharing
symbols globally, and in fact, most modules only export the symbols
"mod_name" and "mod_main". However, if a broker module itself wants
to load plugins, and _also_ provide some symbols for those plugins
to use, then use of RTLD_LOCAL prevents this.

Since most modules only export the "mod_{main,name}" symbols
mentioned above, there is likely little current risk in using
RTLD_GLOBAL. Therefore, load all broker modules with this flag to
enable the plugin-loading module use case by default.

If RTLD_GLOBAL for all modules becomes a problem in the future, then
RTLD_LOCAL should be reinstated as the default, with an option for
modules to *request* that they be loaded with RTLD_GLOBAL. (This can
be done by reloading the DSO with RTLD_NOLOAD | RTLD_GLOBAL. This
is used to promote the flags on a shared object. See the dlopen(3)
manual page for more information).