configure: add switch to allow flux to be built without python
Problem: Build errors occur when using the Julia packaging tool
Yggdrasil to build Flux and package libflux.so as part of a Julia
package. These errors occur due to python versioning and the python
package dependencies.

Solution: Add a configure flag that disables the python
dependencies. This obviously produces an otherwise broken Flux build (no
working front-end utilities or validator), but it is useful for
producing a libflux.so.

Closes #3420