libflux: test: fix flux_plugin_call() tests
Problem: Tests for flux_plugin_call() check for rc == 0 for success,
but flux_plugin_call() should have a way to detect if a callback
was called, so may return >= 0 on success.

Fix all `flux_plugin_call()` checks in libflux/test/plugin.c to check
for rc >= 0 on success instead of rc == 0.