libflux: only return 0 from flux_plugin_call() if no callback
Problem: There is no way to determine if flux_plugin_call() found
a matching callback and called it, vs returned 0 because no callback
was registered for the given name.

Return 1 from flux_plugin_call() when a callback was called with a
return value of >= 0. A return code of 0 now means that no callback
was found (so the function did nothing), while < 0 indicates an
error.