python: ensure reactor_depth is initialized per-thread
Problem: The Python Flux handle class attempts to instantiate a
global per-thread "reactor_depth" parameter using threading.local(),
but this parameter is currently only initialized to 0 in the main
thread, when the class itself is first compiled by the process.
This leads to an exception when a thread tries to use reactor_run().

Ensure reactor depth is initialized for any thread that calls
flux.Flux() by checking for "reactor_depth" in self.tls.__dict__.
If this key is not set in the dict, then intialize *both*
self.tls.reactor_depth and self.tls.exception.