job-manager: add priorization module
Problem: The code for dealing with job priority and urgency
is sprinkled through the job manager code base. In several places
urgency is used in place of priority, while in others the priority
is explicitly set to the urgency.

Refine and abstract most job priority handling to a new prioritize
module. The main useful function in prioritize.c sets the priority
of a single job and takes the appriopriate action -- generating
a "priority" event if the priority was changed, and notifying the
scheduler's prioritize RPC if necessary.

This function also ensures that priority is overridden by urgency
when job->urgency is set to the *administrative* values of HOLD
and EXPEDITE, which always override the computed priority. (In the
current code, this isn't the case since job->priority == urgency.
However, this paves the way for plugin-based priority calculations
in the future.)

Additionally, fix the use of job->urgency in several places in the
code where job->priority should be used instead.

Fixes #3452