job-manager: add jobtap plugin interface
Problem: There is no facility in the job manager to allow configuration
and/or custom computation of job priorities. This will be necessary
for advanced priority use cases like fairshare scheduling, etc.

Add a simple plugin architecture to the job manaager which allows a plugin
to tap into job state transitions as well as provide a job priority when
the job manager needs to calculate one. This is called the "jobtap"
interface.

The jobtap interface is built using the standard flux_plugin_t
interface, and therefore a jobtap plugin need only define a single
callback:

  int flux_plugin_init (flux_plugin_t *p);

from which the plugin should add handlers for the defined jobtap
topic strings, which currently include the following:

  - job.state.[state]: on transition to job state 'state'

  - job.new: when a new job appears in the job manager

  - job.priority.get: when the job manager wants to recalculate
    priority of a job.

Plugin callbacks are of the standard Flux plugin form:

  int callback (flux_plugin_t *p,
                const char *topic,
                flux_plugin_arg_t *args,
                void *arg);

The jobtap interface provides information for the job for which it
is being called via the flux_plugin_arg_t, e.g. via something like:

   flux_plugin_arg_pack (args,
                         FLUX_PLUGIN_ARG_IN,
                         "{s:O s:I s:i s:i s:i s:I s:f}",
                         "jobspec", job->jobspec_redacted,
                         "id", job->id,
                         "userid", job->userid,
                         "urgency", job->urgency,
                         "state", job->state,
                         "prev_state", old_state,
                         "priority", job->priority,
                         "t_submit", job->t_submit);

which can be optionally unpacked via flux_plugin_arg_unpack() in the
callback.