job-manager: jobtap: reprioritize jobs on plugin load
Problem: When a new jobtap plugin is loaded, the plugin may need
to recalculate the priorities of all jobs in PRIORITY or SCHED
states. For instance, jobs may be stuck in the PRIORITY state due
to a previous plugin's inability to provide a priority. If jobs
were not reprioritized on plugin load, then these jobs would stay
in PRIORITY state forever.

Call the job.new callback on all active jobs to inform the loaded
plugin of all active jobs (if necessary), then call reprioritize_all()
so that the plugin has a chance to update priorities.