sched-simple: improve efficiency of prioritize_cb
Problem: In prioritize_cb(), sched-simple reinserts each reprioritized
job back into the queue. This is inefficient for long lists of jobs, and
will slow job throughput when reprioritization in unlimited mode is
frequent with lots of pending jobs.

Re-sort the entire list of jobs after the priority of all jobs has been
updated when more than 3 jobs are being reprioritized