libschedutil: remove sched.ready mode "single"
Per RFC27 change, the mode "single" has been replaced with a mode
"limited" with a "limit" of 1.

Update schedutil_ready() to take the a mode format of "limited=N",
which represents mode "limited" with a "limit" of N.  Update
"sched.ready" rpc call to pass "limit" if necessary.

Update job-manager and sched-simple appropriately for change.

Fixes #2070